updated types setup / refactored some files
