--- jsdoc.conf.json
@@ -1,57 +1,0 @@
-{
-  "tags": {
-    "allowUnknownTags": true,
-    "dictionaries": [
-      "jsdoc",
-      "closure"
-    ]
-  },
-  "source": {
-    "include": [
-      "src"
-    ],
-    "includePattern": "\\.(jsx|js|ts|tsx)$",
-    "excludePattern": "(^|\\/|\\\\)_"
-  },
-  "plugins": [
-    "plugins/markdown",
-    "node_modules/better-docs/typescript"
-  ],
-  "typescript": {
-    "moduleRoot": "src"
-  },
-  "markdown": {
-    "hardwrap": false,
-    "idInHeadings": true
-  },
-  "templates": {
-    "cleverLinks": false,
-    "monospaceLinks": false,
-    "better-docs": {
-      "name": "Interactive Animated Blocks",
-      "title": "Interactive Animated Blocks",
-      "hideGenerator": true
-    }
-  },
-  "opts": {
-    "encoding": "utf8",
-    "destination": "docs",
-    "template": "./node_modules/better-docs",
-    "readme": "./readme.txt",
-    "recurse": true,
-    "sort": true,
-    "theme_opts": {
-      "menu": [
-        {
-          "title": "Interactive Animated Blocks",
-          "link": "https://github.com/erikyo/ssc/"
-        }
-      ]
-    },
-    "allowUnknownTags": true,
-    "dictionaries": [
-      "jsdoc",
-      "closure"
-    ]
-  }
-}
