+++ tsconfig.json
@@ -1,44 +1,41 @@
 {
   "compilerOptions": {
     "target": "ESNext",
-    "module": "esnext",
+    "module": "ESNext",
     "jsx": "react-jsx",
-
     // Search under node_modules for non-relative imports.
     "moduleResolution": "node",
     "sourceMap": true,
-
     // Enable the strictest settings like strictNullChecks & noImplicitAny.
     "strict": true,
-
     "allowSyntheticDefaultImports": true,
     "skipLibCheck": true,
     "strictNullChecks": true,
     "resolveJsonModule": true,
     "forceConsistentCasingInFileNames": true,
-    // Import non-ES modules as default imports."esModuleInterop": true,
+    // Import non-ES modules as default imports.
+    "esModuleInterop": true,
 
     // Tells TypeScript to read JS files, as
     // normally they are ignored as source files
     "allowJs": true,
-
     // Generate d.ts files
-    "declaration": true,
+    "declaration": false,
     "declarationMap": true,
-    "declarationDir": "build/types",
-    "emitDeclarationOnly": true,
+    "emitDeclarationOnly": false,
     "pretty": true,
-
     // Types should go into this directory.
     // Removing this would place the .d.ts files
     // next to the .js files
-    "outDir": "build",
-    "types": [
-      "puppeteer",
-      "jest-environment-puppeteer",
-      "expect-puppeteer"
-    ]
+    "outDir": "./build",
+    "typeRoots": ["./src", "node_modules/@types"]
   },
-  "include": ["src/**/*", "block.json"],
-  "exclude": ["build","coverage"]
+  "include": [
+    "src/**/*"
+  ],
+  "exclude": [
+    "build",
+    "coverage",
+    "node_modules"
+  ]
 }
