--- tsconfig.types.json
@@ -1,10 +1,0 @@
-{
-  "extends": "./tsconfig",
-  "compilerOptions": {
-    "target": "ESNext",
-    "module": "ESNext",
-    "jsx": "react-jsx",
-    "noEmit": false
-  },
-  "exclude": ["**/*.test.ts"]
-}
